+++ example/helper/MbEncodingHelper.php
@@ -3,6 +3,7 @@
 require __DIR__.'/../../vendor/autoload.php';
 
 use LDL\Framework\Helper\MbEncodingHelper;
+use LDL\Framework\Base\Exception\InvalidArgumentException;
 
 echo "UTF-8 is valid?\n";
 echo MbEncodingHelper::isValid('UTF-8') ? 'Yes' : 'No';
@@ -10,7 +11,7 @@ echo MbEncodingHelper::isValid('UTF-8') ? 'Yes' : 'No';
 try{
     echo "\n\nValidate non-existent encoding UTF-22\n";
     MbEncodingHelper::validate('UTF-22');
-}catch(\InvalidArgumentException $e){
+}catch(InvalidArgumentException $e){
     echo $e->getMessage()."\n";
 }