+++ src/LDL/Framework/Base/Collection/Resolver/Collection/CustomResolverCollection.php
@@ -2,19 +2,20 @@
 
 namespace LDL\Framework\Base\Collection\Resolver\Collection;
 
-use LDL\Framework\Base\Collection\Contracts\CollectionInterface;
-use LDL\Framework\Base\Collection\Resolver\Contracts\CustomResolverInterface;
-use LDL\Framework\Base\Collection\Resolver\Key\IntegerKeyResolver;
+use LDL\Framework\Helper\ArrayHelper\ArrayHelper;
 use LDL\Framework\Base\Collection\Traits\AppendManyTrait;
-use LDL\Framework\Base\Collection\Traits\BeforeAppendInterfaceTrait;
-use LDL\Framework\Base\Collection\Traits\BeforeRemoveInterfaceTrait;
+use LDL\Framework\Base\Exception\InvalidArgumentException;
+use LDL\Framework\Base\Traits\LockableObjectInterfaceTrait;
+use LDL\Framework\Base\Collection\Contracts\CollectionInterface;
+use LDL\Framework\Base\Collection\Resolver\Key\IntegerKeyResolver;
 use LDL\Framework\Base\Collection\Traits\CollectionInterfaceTrait;
 use LDL\Framework\Base\Collection\Traits\LockAppendInterfaceTrait;
 use LDL\Framework\Base\Collection\Traits\LockRemoveInterfaceTrait;
 use LDL\Framework\Base\Collection\Traits\RemoveByKeyInterfaceTrait;
-use LDL\Framework\Base\Traits\LockableObjectInterfaceTrait;
-use LDL\Framework\Helper\ArrayHelper\ArrayHelper;
 use LDL\Framework\Helper\ArrayHelper\Exception\InvalidKeyException;
+use LDL\Framework\Base\Collection\Traits\BeforeAppendInterfaceTrait;
+use LDL\Framework\Base\Collection\Traits\BeforeRemoveInterfaceTrait;
+use LDL\Framework\Base\Collection\Resolver\Contracts\CustomResolverInterface;
 
 class CustomResolverCollection implements CustomResolverCollectionInterface
 {
@@ -53,7 +54,7 @@ class CustomResolverCollection implements CustomResolverCollectionInterface
                 is_object($item) ? get_class($item) : gettype($item)
             );
 
-            throw new \InvalidArgumentException($msg);
+            throw new InvalidArgumentException($msg);
         }
 
         $this->setItem($item, $this->keyResolver->resolveCustom($this, $key, $item));