+++ src/LDL/Framework/Base/Collection/Resolver/Key/StringKeyResolver.php
@@ -2,6 +2,7 @@
 
 namespace LDL\Framework\Base\Collection\Resolver\Key;
 
+use LDL\Framework\Base\Exception\LogicException;
 use LDL\Framework\Base\Collection\Contracts\CollectionInterface;
 use LDL\Framework\Base\Collection\Resolver\Contracts\FullResolverInterface;
 
@@ -57,6 +58,6 @@ class StringKeyResolver implements FullResolverInterface
             }
         }
 
-        throw new \LogicException("Could not resolve a suitable key for $key");
+        throw new LogicException("Could not resolve a suitable key for $key");
     }
 }